pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Multi-agent codebase intelligence system for rapid FDE onboarding"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pydantic>=2.0",
    "networkx>=3.0",
    "tree-sitter>=0.21,<0.22",
    "tree-sitter-languages>=1.0",
    "sqlglot>=20.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "gitpython>=3.1",
    "click>=8.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
cartographer = "src.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
